requirements.txt
alembic==1.5.2
Brotli==1.0.9
click==7.1.2
cycler==0.10.0
dash==1.19.0
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.8.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
Pillow==8.1.0
plotly==4.14.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
redis==3.5.3
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.22
visitor==0.1.3
Werkzeug==1.0.1
